requirements.txt

fastapi
uvicorn
redis
psycopg2-binary
